pyproject.toml
[tool.poetry]
name = "apache-airflow-testing"
version = "0.1.0"
description = ""
authors = ["Stephen Dillon <yodareloaded@gmail.com>"]
readme = "README.md"
packages = [{ include = "dags", from = "." }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.31.0"
apache-airflow = "^2.9.3"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-httpserver = "^1.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
